Unsignify lengths and addresses in chip functions and structs
Push those changes forward where needed to prevent new sign
conversion warnings where possible.

Acked-by: Carl-Daniel Hailfinger <c-d.hailfinger.devel.2006@gmx.net>

git-svn-id: svn://flashrom.org/flashrom/trunk@1470 2b7e53f0-3cfb-0310-b3e9-8179ed1497e1

+++ linux_spi.c
@@ -36,10 +36,10 @@ static int fd = -1;
 static int linux_spi_shutdown(void *data);
 static int linux_spi_send_command(unsigned int writecnt, unsigned int readcnt,
 			const unsigned char *txbuf, unsigned char *rxbuf);
-static int linux_spi_read(struct flashchip *flash, uint8_t *buf, int start,
-			  int len);
+static int linux_spi_read(struct flashchip *flash, uint8_t *buf,
+			  unsigned int start, unsigned int len);
 static int linux_spi_write_256(struct flashchip *flash, uint8_t *buf,
-			       int start, int len);
+			       unsigned int start, unsigned int len);
 
 static const struct spi_programmer spi_programmer_linux = {
 	.type		= SPI_CONTROLLER_LINUX,
@@ -131,14 +131,14 @@ static int linux_spi_send_command(unsigned int writecnt, unsigned int readcnt,
 	return 0;
 }
 
-static int linux_spi_read(struct flashchip *flash, uint8_t *buf, int start,
-			  int len)
+static int linux_spi_read(struct flashchip *flash, uint8_t *buf,
+			  unsigned int start, unsigned int len)
 {
-	return spi_read_chunked(flash, buf, start, len, getpagesize());
+	return spi_read_chunked(flash, buf, start, len, (unsigned)getpagesize());
 }
 
 static int linux_spi_write_256(struct flashchip *flash, uint8_t *buf,
-			       int start, int len)
+			       unsigned int start, unsigned int len)
 {
-	return spi_write_chunked(flash, buf, start, len, getpagesize() - 4);
+	return spi_write_chunked(flash, buf, start, len, ((unsigned)getpagesize()) - 4);
 }
